Reorganize code layout.
Instead of one big directory for all the things, add a subdirectory for
all the traits. I feel this keeps the library a bit more structured and
understandable.
