+++ src/Date.php
@@ -23,14 +23,14 @@ use DateTimeZone;
  */
 class Date extends DateTimeImmutable implements ChronosInterface
 {
-    use ComparisonTrait;
-    use DifferenceTrait;
-    use FactoryTrait;
-    use FormattingTrait;
-    use MagicPropertyTrait;
-    use ModifierTrait;
-    use RelativeKeywordTrait;
-    use TestingAidTrait;
+    use Traits\ComparisonTrait;
+    use Traits\DifferenceTrait;
+    use Traits\FactoryTrait;
+    use Traits\FormattingTrait;
+    use Traits\MagicPropertyTrait;
+    use Traits\ModifierTrait;
+    use Traits\RelativeKeywordTrait;
+    use Traits\TestingAidTrait;
 
     /**
      * Format to use for __toString method when type juggling occurs.
