+++ src/Traits/ComparisonTrait.php
@@ -10,9 +10,10 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
 use DateTime;
+use Cake\Chronos\ChronosInterface;
 
 /**
  * Provides various comparison operator methods for datetime objects.
