+++ src/Traits/CopyTrait.php
@@ -10,7 +10,7 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
 /**
  * Provides methods for copying datetime objects.
