+++ src/Traits/DifferenceTrait.php
@@ -10,8 +10,10 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
+use Cake\Chronos\ChronosInterface;
+use Cake\Chronos\ChronosInterval;
 use DatePeriod;
 use DateTimeInterface;
