+++ src/Traits/FormattingTrait.php
@@ -10,8 +10,9 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
+use Cake\Chronos\ChronosInterface;
 use DateTime;
 
 /**
