+++ src/Traits/MagicPropertyTrait.php
@@ -11,8 +11,9 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
+use Cake\Chronos\ChronosInterface;
 use InvalidArgumentException;
 
 /**
