+++ src/Traits/RelativeKeywordTrait.php
@@ -11,7 +11,7 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
 
 /**
  * Provides methods for testing if strings contain relative keywords.
