+++ src/Traits/TestingAidTrait.php
@@ -11,7 +11,9 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Chronos;
+namespace Cake\Chronos\Traits;
+
+use Cake\Chronos\ChronosInterface;
 
 /**
  * Provides methods for setting a 'test' now. This lets you
